added conditional rendering for offer-page component
