+++ src/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './components/app/app';
-import { Setting } from './data';
+import { Setting } from './mocks/mock-data';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
