+++ src/mocks/mock-util.ts
@@ -1,0 +1,3 @@
+import { AuthorizationStatus } from '../data';
+
+export const isUserLogged = (logged: AuthorizationStatus) => logged === AuthorizationStatus.Auth;
